+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 2020 01 10.01
 
+    - Fix for git#22, Preserve function signature on a single line. An
+      unwanted line break was being introduced when a closing signature paren
+      followed a closing do brace.
+
     - Fix RT#132059, the -dac parameter was not working and caused an error exit
 
     - When -utf8 is used, any error output is encoded as utf8